+++ libgcompat/auxv.c
@@ -1,0 +1,6 @@
+#include <sys/auxv.h> /* getauxval */
+
+unsigned long __getauxval(unsigned long value)
+{
+	return getauxval(value);
+}